+++ app/src/main/java/org/simple/clinic/summary/addphone/AddPhoneNumberUiRender.kt
@@ -3,7 +3,6 @@ package org.simple.clinic.summary.addphone
 import org.simple.clinic.mobius.ViewRenderer
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.Blank
-import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.LengthTooLong
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.LengthTooShort
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.ValidNumber
 
@@ -16,7 +15,6 @@ class AddPhoneNumberUiRender(private val ui: AddPhoneNumberUi) : ViewRenderer<Ad
 
   private fun renderValidationErrors(validationResult: Result) {
     when (validationResult) {
-      is LengthTooLong -> ui.showPhoneNumberTooLongError(validationResult.maximumRequiredNumberLength)
       is LengthTooShort -> ui.showPhoneNumberTooShortError(validationResult.minimumAllowedNumberLength)
       Blank -> ui.showPhoneNumberBlank()
       ValidNumber -> ui.clearPhoneNumberError()