+++ app/src/main/java/org/simple/clinic/summary/updatephone/UpdatePhoneNumberDialog.kt
@@ -174,10 +174,6 @@ class UpdatePhoneNumberDialog : AppCompatDialogFragment(), UpdatePhoneNumberDial
     phoneInputLayout.error = getString(R.string.patientsummary_updatephone_error_phonenumber_length_less, minimumAllowedNumberLength.toString())
   }
 
-  override fun showPhoneNumberTooLongError(maximumRequiredNumberLength: Int) {
-    phoneInputLayout.error = getString(R.string.patientsummary_updatephone_error_phonenumber_length_more, maximumRequiredNumberLength.toString())
-  }
-
   override fun preFillPhoneNumber(number: String) {
     numberEditText.setTextAndCursor(number)
   }