+++ app/src/test/java/org/simple/clinic/patient/OngoingNewPatientEntryTest.kt
@@ -12,7 +12,6 @@ import org.simple.clinic.patient.OngoingNewPatientEntry.PersonalDetails
 import org.simple.clinic.patient.OngoingNewPatientEntry.PhoneNumber
 import org.simple.clinic.registration.phone.PhoneNumberValidator
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.ValidNumber
-import org.simple.clinic.registration.phone.PhoneNumberValidator.Type.LANDLINE_OR_MOBILE
 import org.simple.clinic.util.TestUserClock
 import org.simple.clinic.widgets.ageanddateofbirth.UserInputAgeValidator
 import org.simple.clinic.widgets.ageanddateofbirth.UserInputDateValidator
@@ -58,7 +57,7 @@ class OngoingNewPatientEntryTest {
     )
 
     val numValidator = mock<PhoneNumberValidator>()
-    whenever(numValidator.validate("", LANDLINE_OR_MOBILE)).thenReturn(phoneValidationResult)
+    whenever(numValidator.validate("")).thenReturn(phoneValidationResult)
 
     assertThat(entry.validationErrors(dobValidator, numValidator, ageValidator)).apply {
       if (isValid) {
@@ -76,7 +75,6 @@ class OngoingNewPatientEntryTest {
         arrayOf(false, "Ashok Kumar", "01-01-1971", "47", "colony", "state", "district", PhoneNumberValidator.Result.LengthTooShort(6)),
         arrayOf(false, "Ashok Kumar", "01/01/1971", null, "colony", "state", "district", PhoneNumberValidator.Result.Blank),
         arrayOf(false, "Ashok Kumar", "01/01/1971", null, "colony", "state", "district", PhoneNumberValidator.Result.LengthTooShort(6)),
-        arrayOf(false, "Ashok Kumar", "01/01/1971", null, "colony", "state", "district", PhoneNumberValidator.Result.LengthTooLong(12)),
         arrayOf(true, "Ashok Kumar", "01/01/1971", null, "colony", "state", "district", PhoneNumberValidator.Result.ValidNumber)
     )
   }
@@ -100,7 +98,7 @@ class OngoingNewPatientEntryTest {
     whenever(mockDobValidator.validate("01/01/3000")).thenReturn(DateIsInFuture)
 
     val mockNumValidator = mock<PhoneNumberValidator>()
-    whenever(mockNumValidator.validate("phone-number", LANDLINE_OR_MOBILE)).thenReturn(ValidNumber)
+    whenever(mockNumValidator.validate("phone-number")).thenReturn(ValidNumber)
 
     val validationErrors = entry.validationErrors(mockDobValidator, mockNumValidator, mockAgeValidator)