+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changes
 
 - Show alpha numeric keyboard for national ID text field in patient entry and edit screen
+- Remove maximum character limit for phone numbers
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen
   - Show hypertension diagnosis and treatment in single card
   - Show separate cards for hypertension and diabetes diagnosis